hotfix: fix modifiers crashing when attempting to assign to a nonparticipating player

+++ src/main/java/com/maximumg9/shadow/abilities/RoleGuess.java
@@ -96,7 +96,7 @@ public class RoleGuess extends Ability {
                             player.role != null &&
                                 player.role.getFaction() == Faction.SHADOW
                     ).count();
-                long nonShadows = (long) getShadow().indirectPlayerManager
+                long nonShadows = getShadow().indirectPlayerManager
                     .getRecentlyOnlinePlayers(getShadow().config.disconnectTime)
                     .stream().filter(
                         (player) -> player.role != null &&
